Little Merges (again, most done already)

+++ src/game/file.c
@@ -37,6 +37,7 @@ s32 save_selected_bond[] = {0,0,0,0};
 //D:8002C520
 save_data D_8002C520 = {0, 0, 0x80, 0x00, 0xFF, 0xFF, DEFAULT_OPTIONS, 0x00, 0x00, 0, 0, 0, 0, 0};
 //D:8002C580
+// Default Save
 save_data D_8002C580 = {0, 0, 0x80, 0x00, 0xFF, 0xFF, DEFAULT_OPTIONS, 0x00, 0x00, 0, 0, 0, 0, 0};
 //D:8002C5E0
 save_data D_8002C5E0 = {0, 0, 0x80, 0x00, 0xFF, 0xFF, DEFAULT_OPTIONS, 0x00, 0x00, 0, 0, 0, 0, 0};
@@ -131,7 +132,7 @@ void copyCurrentEEPROMtoStack(void)
   fileCopySaveIfSelectedBondDifferent(selected_folder_num);
 }
 
-u8 getSelectedFolderBond(void)
+s32 getSelectedFolderBond(void)
 {
   return fileGetBondForFolder(selected_folder_num);
 }
